requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
bibtexparser==1.4.3
certifi==2025.11.12
charset-normalizer==3.4.4
chex==0.1.91
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
dm-haiku==0.0.15
e3nn-jax==0.20.8
einops==0.8.1
etils==1.13.0
executing==2.2.1
flax==0.12.1
fonttools==4.60.1
freud-analysis==3.4.0
frozendict==2.4.7
fsspec==2025.10.0
humanize==4.14.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jax==0.8.1
jax-md==0.2.27
jaxlib==0.8.1
jedi==0.19.2
jmp==0.0.4
joblib==1.5.2
jraph==0.0.6.dev0
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_collections==1.1.0
ml_dtypes==0.5.4
monty==2025.3.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
narwhals==2.12.0
nest_asyncio==1.6.0
networkx==3.5
numpy==2.3.5
opt_einsum==3.4.0
optax==0.2.6
orbax-checkpoint==0.11.28
orjson==3.11.4
packaging==25.0
palettable==3.3.3
pandas==2.3.3
parsnip-cif==0.4.1
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
pip==25.3
platformdirs==4.5.0
plotly==6.5.0
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pymatgen==2025.10.7
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.2.0
rowan==1.3.2
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
scipy==1.16.3
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
spglib==2.6.0
stack_data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorstore==0.1.79
toolz==1.1.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
treescope==0.1.10
typing_extensions==4.15.0
tzdata==2025.2
uncertainties==3.2.3
urllib3==2.5.0
wcwidth==0.2.14
wheel==0.45.1
zipp==3.23.0
